backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "the-unfinished-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "python-multipart>=0.0.9",
    "anthropic>=0.28.0",
    "python-docx>=1.1.0",
    "pypdf>=4.2.0",
    "httpx>=0.27.0",
    "nltk>=3.8.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
where = ["."]
